fix: Fullscreen button margin below PowerBI reports
A different strategy to positioning the fullscreen button is used.
The fullscreen button is not outside of the PowerBI report anymore,
which means that default margins are correct and the fullscreen button
cannot be "above" below paragraphs.

The edge case is now for when there are internalLinkParagraph nodes
directly after the report. In this case, we lift up those nodes so
that they appear at the same level as the fullscreen button.